scheduler/requirements.txt
kubernetes>=28.1.0
pyyaml>=6.0.1

